+++ src/supermodels/converters/__init__.py
@@ -1,0 +1,1 @@
+# ~/supermodels/src/supermodels/converters/__init__.py